pyproject.toml
[project]
name = "nno0obb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.12"
dependencies = [
    "gitpython>=3.1.45",
    "tqdm>=4.67.1",
]
